requirements.txt
cachetools
matplotlib
numpy
numba
opencv-python
perlin-numpy
pygame-ce
pyinstaller
scipy
icosphere